rename dispatcher file
